refactor: decompose InferenceService into ModelLifecycleCoordinator + GenerationCoordinator (#279)
Split the 972-LOC InferenceService god-object into focused internal types
while preserving the identical public API. InferenceService becomes a thin
facade that dispatches to:

- ModelLifecycleCoordinator (471 LOC): backend registry, model load/unload,
  LoadRequestToken lifecycle, progress reporting, memory gating, capability
  and compatibility queries, tokenizer vending.

- GenerationCoordinator (302 LOC): generation queue (enqueue/cancel/stop),
  drainQueue with its critical defer-block auto-drain, low-level generate()
  prompt formatting, session-scoped request filtering.

The two coordinators communicate via GenerationContextProvider, an internal
protocol that InferenceService conforms to, providing the generation
coordinator read access to the backend, model-loaded state, and prompt
template without a direct dependency on the lifecycle coordinator.

Also adds 23 new tests:
- 13 regression lock tests covering previously untested public API
  (resetConversation, registeredBackendSnapshot, compatibility, lastTokenUsage,
  generationDidFinish no-op, declareSupport accumulation)
- 6 observation propagation tests using withObservationTracking
- 4 facade integration tests verifying cross-coordinator dispatch

Fixes the timing-sensitive test_enqueue_priority_threeLevel_ordering test
to consume streams deterministically instead of relying on Task.yield().

Closes #279
