+++ Tests/BaseChatInferenceTests/SilentCatchAuditTest.swift
@@ -43,9 +43,9 @@ final class SilentCatchAuditTest: XCTestCase {
         "BaseChatInference/Models/ModelInfo.swift:if let metadata = try? GGUFMetadataReader.readMetadata(from: url) {",
         "BaseChatInference/Models/ModelInfo.swift:guard let contents = try? fileManager.contentsOfDirectory(",
         "BaseChatInference/Models/ModelInfo.swift:let values = try? fileURL.resourceValues(forKeys: [.fileSizeKey, .isRegularFileKey])",
-        "BaseChatInference/Services/BackgroundDownloadManager.swift:guard let handle = try? FileHandle(forReadingFrom: fileURL) else {",
-        "BaseChatInference/Services/BackgroundDownloadManager.swift:guard let headerData = try? handle.read(upToCount: 4), headerData.count == 4 else {",
         "BaseChatInference/Services/BackgroundDownloadManager.swift:try? FileManager.default.removeItem(at: tempURL)",
+        "BaseChatInference/Services/DownloadFileValidator.swift:guard let handle = try? FileHandle(forReadingFrom: fileURL) else {",
+        "BaseChatInference/Services/DownloadFileValidator.swift:guard let headerData = try? handle.read(upToCount: 4), headerData.count == 4 else {",
         "BaseChatInference/Services/GGUFMetadataReader.swift:guard let handle = try? FileHandle(forReadingFrom: url) else { return false }",
         "BaseChatInference/Services/ModelStorageService.swift:guard let contents = try? fileManager.contentsOfDirectory(",
 
@@ -58,7 +58,6 @@ final class SilentCatchAuditTest: XCTestCase {
 
         // BaseChatUI — Task.sleep cancellation is intentionally ignored;
         // parser/rendering fallbacks are benign optional conversions.
-        "BaseChatUI/ViewModels/ChatViewModel+Generation.swift:guard let regex = try? NSRegularExpression(pattern: pattern, options: []) else {",
         "BaseChatUI/ViewModels/ModelManagementViewModel.swift:try? await Task.sleep(for: .milliseconds(500))",
         "BaseChatUI/Views/Chat/AssistantMarkdownView.swift:if let parsed = try? AttributedString(",
         "BaseChatUI/Views/Chat/TypingIndicatorView.swift:try? await Task.sleep(for: .milliseconds(400))",
@@ -71,6 +70,8 @@ final class SilentCatchAuditTest: XCTestCase {
         "BaseChatTestSupport/PerceivedLatencyBackend.swift:try? await Task.sleep(for: delay)",
         "BaseChatTestSupport/ChaosBackend.swift:try? await Task.sleep(for: delay)",
         "BaseChatTestSupport/ChaosBackend.swift:try? await Task.sleep(for: stallDuration)",
+        "BaseChatTestSupport/HardwareRequirements.swift:let configData = try? Data(contentsOf: configURL),",
+        "BaseChatTestSupport/HardwareRequirements.swift:let json = try? JSONSerialization.jsonObject(with: configData) as? [String: Any],",
         "BaseChatTestSupport/HardwareRequirements.swift:let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any],",
         "BaseChatTestSupport/HardwareRequirements.swift:if let containers = try? fm.contentsOfDirectory(",
         "BaseChatTestSupport/HardwareRequirements.swift:guard let contents = try? fm.contentsOfDirectory(",
